Fix return type of hs_popcntX
All the hs_popcntX functions should return an StgWord in accordance with
the primop types.
I'm not sure whether the argument should also become an StgWord
(except for the 64-bit versions), the primops have type Word# -> Word#,
but we have e.g. StgWord hs_popcnt8(StgWord8 x).
StgWord8 is an unsigned char (usually, at least), but I think arguments
are passed at least word-sized to C-functions, so it probably works.
For the moment it works and passes tests, I'll ask tibbe to be sure.